Upgrade apollo-server-micro, move Prisma from lib to schema to fix dev server issue
