+++ package.json
@@ -23,7 +23,7 @@
     "prepare": "yarn generate",
     "husky-install": "husky install",
     "clean-modules": "cli-confirm \"Are you sure?\" && find . -name \"node_modules\" -type d -prune -exec rm -rf '{}' +",
-    "generate": "yarn workspace @acter/schema generate",
+    "generate": "yarn workspace @acter/schema build",
     "db:migrate": "yarn workspace @acter/schema db:migrate",
     "db:migrate:run": "yarn workspace @acter/schema db:migrate:run",
     "db:reset": "yarn workspace @acter/schema db:reset",
