+++ packages/lib/acter/get-or-create-acter-from-db.ts
@@ -1,4 +1,4 @@
-import prisma from '@acter/lib/prisma'
+import prisma from '@acter/schema/prisma'
 import { getActerTypeFromDB } from '@acter/lib/acter-types/get-acter-type-from-db'
 import { getUniqueActerSlugFromDB } from '@acter/lib/acter/get-unique-acter-slug-from-db'
 import { ActerTypes } from '@acter/lib/constants'
