+++ packages/web/package.json
@@ -17,10 +17,11 @@
     "@material-ui/core": "*",
     "@next/bundle-analyzer": "^11.0.1",
     "@sentry/nextjs": "^6.8.0",
-    "apollo-server-micro": "^2.25.2",
+    "apollo-server-micro": "^3.1.2",
     "aws-sdk": "^2.940.0",
     "env-cmd": "^10.1.0",
     "lodash": "^4.17.21",
+    "micro": "^9.3.4",
     "next": "*",
     "next-compose-plugins": "^2.2.1",
     "next-graphql-loader": "^1.0.1",
